implement dismiss for orchestrated enhance selfie capture
