+++ Example/SmileID/Home/HomeView.swift
@@ -139,109 +139,6 @@ struct HomeView: View {
     }
 }
 
-// We need to define a separate proxy delegate because it's the same protocol for both Enrollment
-// and Authentication. However, since the result is still processing, the result parameter is not
-// yet populated (which is what contains the jobType). On Enroll, we need to perform a different
-// action (namely, save userId to clipboard)
-class SmartSelfieEnrollmentDelegate: SmartSelfieResultDelegate {
-    let userId: String
-    let onEnrollmentSuccess: (
-        _ userId: String,
-        _ selfieFile: URL,
-        _ livenessImages: [URL],
-        _ apiResponse: SmartSelfieResponse?
-    ) -> Void
-    let onError: (Error) -> Void
-
-    init(
-        userId: String,
-        onEnrollmentSuccess: @escaping (
-            _: String,
-            _: URL,
-            _: [URL],
-            _: SmartSelfieResponse?
-        ) -> Void,
-        onError: @escaping (
-            Error
-        ) -> Void
-    ) {
-        self.userId = userId
-        self.onEnrollmentSuccess = onEnrollmentSuccess
-        self.onError = onError
-    }
-
-    func didSucceed(
-        selfieImage: URL,
-        livenessImages: [URL],
-        apiResponse: SmartSelfieResponse?
-    ) {
-        onEnrollmentSuccess(userId, selfieImage, livenessImages, apiResponse)
-    }
-
-    func didError(error: Error) {
-        onError(error)
-    }
-}
-
-private struct SmartSelfieAuthWithUserIdEntry: View {
-    let initialUserId: String
-    let useStrictMode: Bool
-    let delegate: SmartSelfieResultDelegate
-    let onDismiss: (() -> Void)
-
-    @State private var userId: String?
-
-    init(
-        initialUserId: String,
-        useStrictMode: Bool = false,
-        delegate: SmartSelfieResultDelegate,
-        onDismiss: @escaping () -> Void
-    ) {
-        self.initialUserId = initialUserId
-        self.delegate = delegate
-        self.useStrictMode = useStrictMode
-        self.onDismiss = onDismiss
-    }
-
-    var body: some View {
-        if let userId {
-            ZStack {
-                if useStrictMode {
-                    SmileID.smartSelfieAuthenticationScreenEnhanced(
-                        config: OrchestratedSelfieCaptureConfig(
-                            userId: userId,
-                            isEnroll: false
-                        ),
-                        delegate: delegate
-                    )
-                } else {
-                    SmileID.smartSelfieAuthenticationScreen(
-                        config: OrchestratedSelfieCaptureConfig(
-                            userId: userId,
-                            isEnroll: false,
-                            allowAgentMode: true
-                        ),
-                        delegate: delegate,
-                        onDismiss: onDismiss
-                    )
-                }
-            }
-            .transition(.move(edge: .trailing))
-        } else {
-            CancellableNavigationView {
-                EnterUserIDView(initialUserId: initialUserId) { userId in
-                    withAnimation {
-                        self.userId = userId
-                    }
-                }
-            } onCancel: {
-                onDismiss()
-            }
-            .transition(.move(edge: .leading))
-        }
-    }
-}
-
 private struct HomeView_Previews: PreviewProvider {
     static var previews: some View {
         let _ = SmileID.initialize(
