+++ Sources/SmileID/Classes/SelfieCapture/View/OrchestratedEnhancedSelfieCaptureScreen.swift
@@ -29,40 +29,45 @@ public struct OrchestratedEnhancedSelfieCaptureScreen: View {
     }
 
     public var body: some View {
-        NavigationView {
-            if showInstructions {
-                LivenessCaptureInstructionsView(
-                    showAttribution: config.showAttribution,
-                    didTapGetStarted: {
-                        showInstructions = false
+        CancellableNavigationView {
+            ZStack {
+                if showInstructions {
+                    LivenessCaptureInstructionsView(
+                        showAttribution: config.showAttribution,
+                        didTapGetStarted: {
+                            showInstructions = false
+                        }
+                    )
+                    .transition(.move(edge: .leading))
+                } else {
+                    ZStack {
+                        if let processingState = viewModel.processingState {
+                            EnhancedSelfieCaptureStatusView(
+                                processingState: processingState,
+                                errorMessage: processingState == .error ? getErrorSubtitle(
+                                    errorMessageRes: viewModel.errorMessageRes,
+                                    errorMessage: viewModel.errorMessage
+                                ) : nil,
+                                selfieImage: viewModel.selfieImage,
+                                showAttribution: config.showAttribution,
+                                didTapCancel: { viewModel.handleCancelSelfieCapture() },
+                                didTapRetry: { viewModel.handleRetry() }
+                            )
+                        } else {
+                            EnhancedSelfieCaptureScreen(
+                                userId: config.userId,
+                                showAttribution: config.showAttribution,
+                                delegate: viewModel,
+                                didTapCancel: { viewModel.handleCancelSelfieCapture() }
+                            )
+                        }
                     }
-                )
-                .transition(.move(edge: .leading))
-            } else {
-                ZStack {
-                    if let processingState = viewModel.processingState {
-                        EnhancedSelfieCaptureStatusView(
-                            processingState: processingState,
-                            errorMessage: processingState == .error ? getErrorSubtitle(
-                                errorMessageRes: viewModel.errorMessageRes,
-                                errorMessage: viewModel.errorMessage
-                            ) : nil,
-                            selfieImage: viewModel.selfieImage,
-                            showAttribution: config.showAttribution,
-                            didTapCancel: { viewModel.handleCancelSelfieCapture() },
-                            didTapRetry: { viewModel.handleRetry() }
-                        )
-                    } else {
-                        EnhancedSelfieCaptureScreen(
-                            userId: config.userId,
-                            showAttribution: config.showAttribution,
-                            delegate: viewModel,
-                            didTapCancel: { viewModel.handleCancelSelfieCapture() }
-                        )
-                    }
+                    .transition(.move(edge: .trailing))
                 }
-                .transition(.move(edge: .trailing))
             }
+        } onCancel: {
+            onDismiss?()
         }
+
     }
 }
